Transaction POST done + refactor Service/DAO of Sectors/Company + fixed tests

+++ src/main/java/com/cs/trader/services/CompanyService.java
@@ -2,6 +2,8 @@ package com.cs.trader.services;
 
 import com.cs.trader.dao.CompanyDao;
 import com.cs.trader.domain.Company;
+import com.cs.trader.domain.Order;
+import com.cs.trader.exceptions.CompanyHasExistingOrdersException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -11,6 +13,12 @@ import java.util.List;
 public class CompanyService {
 	@Autowired
     private CompanyDao companyDao;
+
+	@Autowired
+	private SectorService sectorService;
+
+	@Autowired
+	private OrderService orderService;
 	
 	public List<Company> findAllCompanies(){
 		return companyDao.findAllCompanies();
@@ -28,19 +36,31 @@ public class CompanyService {
 		return companyDao.findAllCompaniesBySectorID(id);
 	}
 
-	public Company validateCompanyByTicker(String ticker){
+	Company validateCompanyByTicker(String ticker){
 		return companyDao.validateTicker(ticker);
 	}
 
 	public int addNewCompany(Company company) {
+		sectorService.findSectorByID(company.getSectorID());
 		return companyDao.addNewCompany(company);
 	}
 
 	public int deleteCompany(int companyID){
-		return companyDao.deleteCompany(companyID);
+		int ordersForCompany = orderService.retrieveOrdersByCompany(findCompanyByID(companyID)).size();
+		if(ordersForCompany == 0) {
+			return companyDao.deleteCompany(companyID);
+		}else {
+			throw new CompanyHasExistingOrdersException("Company ID["+companyID+"] has some valid orders, cannot perform delete operation!");
+		}
 	}
 
-	public Company updateCompany(int id, Company company) {
-		return companyDao.updateCompany(id, company);
+	public Company updateCompany(int id, Company newComp) {
+		Company oldComp = findCompanyByID(id);
+
+		String ticker = newComp.getTicker() == null ? oldComp.getTicker() : newComp.getTicker();
+		String companyName = newComp.getCompanyName() == null ? oldComp.getCompanyName() : newComp.getCompanyName();
+		int sectorID = newComp.getSectorID() == 0 ? oldComp.getSectorID() : sectorService.findSectorByID(newComp.getSectorID()).getSectorID();
+
+		return companyDao.updateCompany(new Company(id, companyName, ticker, sectorID));
 	}
 }
